+++ src/drafts/draft-ietf-cose-merkle-tree-proofs/index.ts
@@ -1,0 +1,3 @@
+import * as receipt from './receipt'
+
+export { receipt }
