+++ src/drafts/draft-ietf-cose-merkle-tree-proofs/receipt/consistency/index.ts
