+++ src/drafts/draft-ietf-cose-merkle-tree-proofs/receipt/get.ts
@@ -1,9 +1,9 @@
-import { decodeFirstSync } from '../../cbor'
+import { decodeFirstSync } from '../../../cbor'
 
-import { CoseSign1Bytes } from "../sign1";
+import { CoseSign1Bytes } from "../../../cose/sign1";
 
-import { draft_headers } from '../../iana/requested/cose';
-import * as cbor from '../../iana/assignments/cbor';
+import { draft_headers } from '../../../iana/requested/cose';
+import * as cbor from '../../../iana/assignments/cbor';
 
 export const get = async (signature: CoseSign1Bytes): Promise<CoseSign1Bytes[]> => {
   const { tag, value } = decodeFirstSync(signature)
