+++ src/drafts/draft-ietf-cose-merkle-tree-proofs/receipt/inclusion/index.ts
