+++ src/drafts/draft-ietf-cose-merkle-tree-proofs/receipt/leaf.ts
